Starting integration test Starting dynamic shared memory
Current bug in add relates to failed alignment when exponent delta is same as number of limbs.  E.g. 2 + 0.1 ends up wrong because the 2 is shifted out.  ShiftLeft seems kind of weird to include for that reason.  If we go back to ShiftRight then we end up with 2 + 0.1 = 2 which is also obviously wrong.