Multiway reference add actually is working now.  No known actual bugs in the reference implementation.
The reference add test will hit some asserts currently because we're not keeping track of precisely how much precision we should be achieving.  The problem arises when we convert an MPIR number into a the GPU-style format, which loses the guard words MPIR uses.  For some numbers these missing values are relevant.  So we need to clean this up.