Start hooking up CUDA multiway add again.
Last time I got this far, I discovered a fundamental bug with the approach after adding some tests, so had to go back to working on the reference implementation.  Now that the reference design is working, back to CUDA.

To recap: The idea with multiway add is to calculate A-B+C and D+E, where we have 5 inputs and 2 outputs.  This calculation corresponds neatly with X^2 - Y^2 + C0_real and 2*X*Y + C0_imaginary.  Simple.

Once this part works we'll have all the pieces for a reference calculation.  Next week is vacation, so the week after will be continuing on the CUDA implementation of multiway add, which I think should work OK.  It won't beat the CPU-only approach but it should be good enough in the context of the calculation of interest, where the multiplies dominate the time spent.