Remove constant memory HDRFloat initialization.
It's unclear whether this approach is better or worse perf.  It was similar before, but I believe constant memory did improve it slightly; thus, this commit may represent a per-pixel perf regression.  But it solves a weird CUDA linker issue.  So let's just roll the dice and see how it goes.