There is something broken with view 30 and it's unclear what it is.
One problem relates to the string to MPIR conversion, which this commit fixes; nevertheless, the "something broken" persists, and that particular cooridinate set is still not working properly.  TBD.

The MPIR function that converts the string to mpf_t leaves it a non-normalized form, wherein the high-order limbs are 0.  In doing so, the function that converts the high-precision mpf_t to HdrFloat implodes and has undefined results.  It uses the x64 bsr instruction to find the first set 0 but that doesn't work here because all the bits are 0, and the result is undefined.  The docs do say that but because this is ameteur hour I failed to read that.