More refactor, hookup custom heap allocator.  Start isolating leaks in the test program. There's a correctness bug in the reference orbit implementation somewhere.  Current approach is (a) fix leaks, (b) run existing tests overnight once leaks fixed, and (c) hopefully find source of ref orbit inaccuracy.
The issue is a race/uninitialized memory or similar.  It's non-deterministic, so it's annoying to guess where it is.  The error is only ever observed after 100K+ iterations of the ref orbit at large limb counts, e.g. 16K limbs, 600K iterations works sometimes, so it's rare.  Maybe one in a million iterations exhibits the issue.