Misc cleanup.
Do full reference orbit equality test between CPU and GPU implementations.  It checks that each HdrFloat entry is identical and that the detected period (if any) matches.

Next we need to hook up the modified bounded-iteration kernel to FractalShark (still busted with this commit).  The objective here is to run a fixed number of iterations repeatedly, so that it's not necessary to allocate huge amounts of memory.  Especially useful if the period is unknown in advance (common case).